Positioned other headings (3).
Applied an alternative strategy, by using div's to encapsulate my child contents so that it is easier to manipulate position without compromising other external content (beyond section/page). Fixed the z-index of the navbar, as it was dissappearing behind the black sections. Applied new font and styling to the band section as well as the general body of text.